Added VCR tests for list_channels method
