+++ lib/slack.rb
@@ -1,39 +1,25 @@
 #!/usr/bin/env ruby
 
-require "dotenv"
-require "httparty"
-require_relative "workspace"
-
-Dotenv.load
-
-SLACK_TOKEN = ENV["SLACK_TOKEN"]
-
 def main
   puts "Welcome to the Ada Slack CLI!"
   workspace = Workspace.new
 
-  # TODO project
+  i = -1
+  while i < 0
+    puts "Your workspace has #{num_channels} channels and #{num_users} users."
+    puts "--------------------"
+    puts "What would you like to do? 'list channels', 'list users', or 'quit'"
+    answer = gets.chomp.downcase
+    if answer == "list channels"
+      # puts workspace.list_channels (method that shows channels)
+    elsif answer == "list users"
+      # puts workspace.list_users
+    elsif answer == "quit"
+      i = 1
+    end
+  end
 
   puts "Thank you for using the Ada Slack CLI"
 end
-
-
-# https://slack.com/api/channels.list?token=xoxb-994132166407-994615665957-EK5sTJjcfCP0o8jQGfelq7f4
- 
-def list_channels
-  url = "https://slack.com/api/channels.list?token=#{SLACK_TOKEN}&pretty=1"
-  response = HTTParty.get(url)
-
-  channels_list = []
-  # looking for channels in response
-  response["channels"].each do |channel|
-    channels_list << channel["name"]
-  end
-
-  return channels_list
-end
-
-p test
-
 
 main if __FILE__ == $PROGRAM_NAME
