+++ test/cassettes/list_channels.yml
@@ -1,0 +1,187 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://slack.com/api/channels.list?pretty=1&token=<SLACK_TOKEN>
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - application/json; charset=utf-8
+      Content-Length:
+      - '654'
+      Connection:
+      - keep-alive
+      Date:
+      - Wed, 11 Mar 2020 22:34:07 GMT
+      Server:
+      - Apache
+      X-Slack-Req-Id:
+      - a2aeede3c2226e70454c8f236a88ec59
+      X-Oauth-Scopes:
+      - channels:read,chat:write,chat:write.customize,chat:write.public,users:read
+      X-Accepted-Oauth-Scopes:
+      - channels:read
+      Access-Control-Expose-Headers:
+      - x-slack-req-id, retry-after
+      X-Slack-Backend:
+      - h
+      X-Content-Type-Options:
+      - nosniff
+      Expires:
+      - Mon, 26 Jul 1997 05:00:00 GMT
+      Cache-Control:
+      - private, no-cache, no-store, must-revalidate
+      X-Xss-Protection:
+      - '0'
+      Vary:
+      - Accept-Encoding
+      Pragma:
+      - no-cache
+      Access-Control-Allow-Headers:
+      - slack-route, x-slack-version-ts, x-b3-traceid, x-b3-spanid, x-b3-parentspanid,
+        x-b3-sampled, x-b3-flags
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Referrer-Policy:
+      - no-referrer
+      Access-Control-Allow-Origin:
+      - "*"
+      X-Via:
+      - haproxy-www-wl4e
+      X-Cache:
+      - Miss from cloudfront
+      Via:
+      - 1.1 0732be5515ffeda639cfe5f22fb04df6.cloudfront.net (CloudFront)
+      X-Amz-Cf-Pop:
+      - SEA19-C2
+      X-Amz-Cf-Id:
+      - aUlRvoP5c1EPVx18rH-iHc--jWiLG1lMDP8aGt80RR9Qk4o9UuTCeQ==
+    body:
+      encoding: ASCII-8BIT
+      string: |
+        {
+            "ok": true,
+            "channels": [
+                {
+                    "id": "CUTE4M96W",
+                    "name": "general",
+                    "is_channel": true,
+                    "created": 1583872630,
+                    "is_archived": false,
+                    "is_general": true,
+                    "unlinked": 0,
+                    "creator": "UV68ES05B",
+                    "name_normalized": "general",
+                    "is_shared": false,
+                    "is_org_shared": false,
+                    "is_member": false,
+                    "is_private": false,
+                    "is_mpim": false,
+                    "members": [
+                        "UUTK13WE6",
+                        "UUW5E7SF3",
+                        "UV68ES05B",
+                        "UV5UMJ4F6"
+                    ],
+                    "topic": {
+                        "value": "Company-wide announcements and work-based matters",
+                        "creator": "UV68ES05B",
+                        "last_set": 1583872630
+                    },
+                    "purpose": {
+                        "value": "This channel is for workspace-wide communication and announcements. All members are in this channel.",
+                        "creator": "UV68ES05B",
+                        "last_set": 1583872630
+                    },
+                    "previous_names": [],
+                    "num_members": 4
+                },
+                {
+                    "id": "CUTE4MTD0",
+                    "name": "api-testing",
+                    "is_channel": true,
+                    "created": 1583872631,
+                    "is_archived": false,
+                    "is_general": false,
+                    "unlinked": 0,
+                    "creator": "UV68ES05B",
+                    "name_normalized": "api-testing",
+                    "is_shared": false,
+                    "is_org_shared": false,
+                    "is_member": false,
+                    "is_private": false,
+                    "is_mpim": false,
+                    "members": [
+                        "UUTK13WE6",
+                        "UUW5E7SF3",
+                        "UV68ES05B",
+                        "UV5UMJ4F6"
+                    ],
+                    "topic": {
+                        "value": "For covid testing",
+                        "creator": "UV68ES05B",
+                        "last_set": 1583958329
+                    },
+                    "purpose": {
+                        "value": "",
+                        "creator": "",
+                        "last_set": 0
+                    },
+                    "previous_names": [],
+                    "num_members": 4
+                },
+                {
+                    "id": "CV83W5SS3",
+                    "name": "random",
+                    "is_channel": true,
+                    "created": 1583872630,
+                    "is_archived": false,
+                    "is_general": false,
+                    "unlinked": 0,
+                    "creator": "UV68ES05B",
+                    "name_normalized": "random",
+                    "is_shared": false,
+                    "is_org_shared": false,
+                    "is_member": false,
+                    "is_private": false,
+                    "is_mpim": false,
+                    "members": [
+                        "UUTK13WE6",
+                        "UUW5E7SF3",
+                        "UV68ES05B",
+                        "UV5UMJ4F6"
+                    ],
+                    "topic": {
+                        "value": "Non-work banter and water cooler conversation",
+                        "creator": "UV68ES05B",
+                        "last_set": 1583872630
+                    },
+                    "purpose": {
+                        "value": "A place for non-work-related flimflam, faffing, hodge-podge or jibber-jabber you'd prefer to keep out of more focused work-related channels.",
+                        "creator": "UV68ES05B",
+                        "last_set": 1583872630
+                    },
+                    "previous_names": [],
+                    "num_members": 4
+                }
+            ],
+            "response_metadata": {
+                "next_cursor": ""
+            }
+        }
+    http_version: null
+  recorded_at: Wed, 11 Mar 2020 22:34:08 GMT
+recorded_with: VCR 5.1.0
